cli - fix for url options not allowing file:// and ftp://
